Commit: Created the project with a handful of features
Explain:
I started building the movies project for the Udacity Android Developers
Nanodegree. SO far, I have created and implimented a recycler view as well
as in the process of adding item intents to redirect from the main page to
the details page. All of the information is currently from my local
project. I want to get everything running before I start to get
information from the API.